Added UI tabs and menu burger
