+++ store/index.ts
@@ -8,12 +8,14 @@ import {
 } from '@reduxjs/toolkit';
 
 /* Local dependencies */
+import getSubscriptionByIdSlice from './slices/subscriptionByIdSlice';
 import userSlice from './slices/userSlice';
 import subscriptionsSlice from './slices/subscriptionsSlice';
 
 const combineReducer = combineReducers({
   userSlice,
   subscriptionsSlice,
+  getSubscriptionByIdSlice,
 });
 
 export function makeStore() {
